modified changes to the DATaskAssignment.cs so as to UpdateTaskPoints depending on the task completed by the user logged in.
made necessary updates in the usp_task_assignment.sql and usp_tasks.sql to reflect the updated points.
updated minpoints to weeklypoints.

+++ SourceCode/WeShare.BusinessLogic/BLGroups.cs
@@ -54,7 +54,7 @@ namespace WeShare.BusinessLogic
             return objDaGroups.DeleteUserFromGroup(groupName, userId);
         }
 
-        public int GetWeeklyPoints(string userId)
+        public double GetWeeklyPoints(string userId)
         {
             return objDaGroups.GetWeeklyPoints(userId);
         }
